blog/requirement.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
crispy-bootstrap5==0.7
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
Django==4.0.4
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==304
pyzmq==23.0.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.2.1.post0
tzdata==2022.1
virtualenv==20.14.1
wcwidth==0.2.5
